+++ src/Numeric/Backprop.hs
@@ -116,7 +116,6 @@ module Numeric.Backprop (
   , Reifies
   ) where
 
-import           Data.Bifunctor
 import           Data.Maybe
 import           Data.Reflection
 import           Data.Type.Index
@@ -201,21 +200,21 @@ backpropN
     => (forall s. Reifies s W => Prod (BVar s) as -> BVar s b)
     -> Tuple as
     -> (b, Tuple as)
-backpropN f = second ($ E.oneFunc) . E.backpropN E.zeroFuncs f
+backpropN = E.backpropN E.zeroFuncs E.oneFunc
 {-# INLINE backpropN #-}
 
 -- | 'backpropN', but allows you to provide the gradient of the "final
 -- result" with respect to the output of your function.  See 'backpropWith'
 -- for more details.
 --
--- Note that argument order changed in v0.2.3.
+-- Note that argument order changed in v0.2.4.
 --
 -- @since 0.2.0.0
 backpropWithN
     :: (Every Backprop as, Known Length as)
     => (forall s. Reifies s W => Prod (BVar s) as -> BVar s b)
     -> Tuple as
-    -> (b, (b -> b) -> Tuple as) -- ^ Takes function giving gradient of final result given the output of function
+    -> (b, b -> Tuple as)
 backpropWithN = E.backpropWithN E.zeroFuncs
 {-# INLINE backpropWithN #-}
 
@@ -231,7 +230,7 @@ backprop
     => (forall s. Reifies s W => BVar s a -> BVar s b)
     -> a
     -> (b, a)
-backprop f = second ($ E.oneFunc) . E.backprop E.zeroFunc f
+backprop = E.backprop E.zeroFunc E.oneFunc
 {-# INLINE backprop #-}
 
 -- | A version of 'backprop' that allows you to specify the gradent of your
@@ -248,14 +247,14 @@ backprop f = second ($ E.oneFunc) . E.backprop E.zeroFunc f
 -- 'backprop' is essentially 'backpropWith' with @'const' 1@ for scalars
 -- and 'Num' instances.
 --
--- Note that argument order changed in v0.2.3
+-- Note that argument order changed in v0.2.4
 --
 -- @since 0.2.0.0
 backpropWith
     :: Backprop a
     => (forall s. Reifies s W => BVar s a -> BVar s b)
     -> a
-    -> (b, (b -> b) -> a) -- ^ Takes function giving gradient of final result given the output of function
+    -> (b, b -> a)
 backpropWith = E.backpropWith E.zeroFunc
 {-# INLINE backpropWith #-}
 
@@ -304,14 +303,14 @@ backprop2
     -> a
     -> b
     -> (c, (a, b))
-backprop2 f x = second ($ E.oneFunc) . E.backprop2 E.zeroFunc E.zeroFunc f x
+backprop2 = E.backprop2 E.zeroFunc E.zeroFunc E.oneFunc
 {-# INLINE backprop2 #-}
 
 -- | 'backprop2', but allows you to provide the gradient of the "final
 -- result" with respect to the output of your function.  See 'backpropWith'
 -- for more details.
 --
--- Note that argument order changed in v0.2.3
+-- Note that argument order changed in v0.2.4
 --
 -- @since 0.2.0.0
 backpropWith2
@@ -319,7 +318,7 @@ backpropWith2
     => (forall s. Reifies s W => BVar s a -> BVar s b -> BVar s c)
     -> a
     -> b
-    -> (c, (c -> c) -> (a, b)) -- ^ Takes function giving gradient of final result given the output of function
+    -> (c, c -> (a, b))
 backpropWith2 = E.backpropWith2 E.zeroFunc E.zeroFunc
 {-# INLINE backpropWith2 #-}