+++ src/Numeric/Backprop/Num.hs
@@ -88,7 +88,6 @@ module Numeric.Backprop.Num (
   , Reifies
   ) where
 
-import           Data.Bifunctor
 import           Data.Maybe
 import           Data.Reflection
 import           Data.Type.Index
@@ -115,7 +114,7 @@ backpropN
     => (forall s. Reifies s W => Prod (BVar s) as -> BVar s b)
     -> Tuple as
     -> (b, Tuple as)
-backpropN f = second ($ E.ofNum) . E.backpropN E.zfNums f
+backpropN = E.backpropN E.zfNums E.ofNum
 {-# INLINE backpropN #-}
 
 -- | 'Numeric.Backprop.backpropWithN', but with 'Num' constraints instead
@@ -123,14 +122,14 @@ backpropN f = second ($ E.ofNum) . E.backpropN E.zfNums f
 --
 -- See 'backpropN' for information on the 'Every' constraint.
 --
--- Note that argument order changed in v0.2.3.
+-- Note that argument order changed in v0.2.4.
 --
 -- @since 0.2.0.0
 backpropWithN
     :: (Every Num as, Known Length as)
     => (forall s. Reifies s W => Prod (BVar s) as -> BVar s b)
     -> Tuple as
-    -> (b, (b -> b) -> Tuple as) -- ^ Takes function giving gradient of final result given the output of function
+    -> (b, b -> Tuple as)
 backpropWithN = E.backpropWithN E.zfNums
 {-# INLINE backpropWithN #-}
 
@@ -144,7 +143,7 @@ backprop
     => (forall s. Reifies s W => BVar s a -> BVar s b)
     -> a
     -> (b, a)
-backprop f = second ($ E.ofNum) . E.backprop E.zfNum f
+backprop = E.backprop E.zfNum E.ofNum
 {-# INLINE backprop #-}
 
 -- | 'Numeric.Backprop.backpropWith', but with 'Num' constraints instead of
@@ -153,14 +152,14 @@ backprop f = second ($ E.ofNum) . E.backprop E.zfNum f
 -- See module documentation for "Numeric.Backprop.Num" for information on
 -- using this with tuples.
 --
--- Note that argument order changed in v0.2.3.
+-- Note that argument order changed in v0.2.4.
 --
 -- @since 0.2.0.0
 backpropWith
     :: Num a
     => (forall s. Reifies s W => BVar s a -> BVar s b)
     -> a
-    -> (b, (b -> b) -> a) -- ^ Takes function giving gradient of final result given the output of function
+    -> (b, b -> a)
 backpropWith = E.backpropWith E.zfNum
 {-# INLINE backpropWith #-}
 
@@ -192,13 +191,13 @@ backprop2
     -> a
     -> b
     -> (c, (a, b))
-backprop2 f x = second ($ E.ofNum) . E.backprop2 E.zfNum E.zfNum f x
+backprop2 = E.backprop2 E.zfNum E.zfNum E.ofNum
 {-# INLINE backprop2 #-}
 
 -- | 'Numeric.Backprop.backpropWith2', but with 'Num' constraints instead of
 -- 'Backprop' constraints.
 --
--- Note that argument order changed in v0.2.3.
+-- Note that argument order changed in v0.2.4.
 --
 -- @since 0.2.0.0
 backpropWith2
@@ -206,7 +205,7 @@ backpropWith2
     => (forall s. Reifies s W => BVar s a -> BVar s b -> BVar s c)
     -> a
     -> b
-    -> (c, (c -> c) -> (a, b)) -- ^ Takes function giving gradient of final result given the output of function
+    -> (c, c -> (a, b)) -- ^ Takes function giving gradient of final result given the output of function
 backpropWith2 = E.backpropWith2 E.zfNum E.zfNum
 {-# INLINE backpropWith2 #-}